requirements.txt
requests
pandas
python-telegram-bot
python-binance
ta>=0.11.0
sqlalchemy
streamlit
plotly
nest_asyncio
